tests: Move sponsorship tests into a new test_sponsorship.
Sponsorship requests for Cloud realms, remote realms, and remote
servers were scattered across three classes in the 10,000-line
test_stripe, between card and invoice upgrade flows.  Collect them in
one place.

The remote classes only need a billing session for the registered
realm or server; the extra users and audit-log pruning that the
remote billing flow classes added only exist for license-ledger
arithmetic, which these tests don't do.

One test exercises a Stripe upgrade, so include the new module when
regenerating Stripe fixtures for the full suite.
